Code cleanup + reward and payment were combined.
more code cleanup is on the way
