+++ pool/difficulty_adjustment.py
@@ -36,7 +36,7 @@ def get_new_difficulty(
 
     # If we don't have enough partials at this difficulty and time between last and
     # 1st partials is below target time, don't update yet
-    if len(recent_partials) < number_of_partials_target and time_taken < time_target :
+    if len(recent_partials) < number_of_partials_target and time_taken < time_target:
         return current_difficulty
 
     # Adjust time_taken if number of partials didn't reach number_of_partials_target
