--- pool/reward/__main__.py
@@ -1,2 +1,0 @@
-from pool.reward.reward_collector_server import main
-main()
